diseño de aviones y fondo, mejoras

+++ Coord.java
@@ -23,7 +23,7 @@ public class Coord {
     // System.out.println( )
     @Override
     public String toString() {
-      return "[" + x + "," + y + "]";
+      return "[" + (int)x + "," + (int)y + "]";
     }
 
     /*
